fix(vue): add suspense for forms with loadercontentfetch
