+++ packages/components-vue/src/types/props/index.ts
@@ -3,3 +3,4 @@ export * from "./modal";
 export * from "./table";
 export * from "./value";
 export * from "./async";
+export * from "./form";
